inference-server/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.36.0
accelerate>=0.25.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pillow>=10.0.0
python-multipart>=0.0.6
